Revert previous refactor and move msg to quotebook crate instead
