+++ quote-book/Cargo.toml
@@ -4,9 +4,6 @@ version = "1.3.0-pre0"
 authors = ["MobileCoin"]
 edition = "2021"
 
-[features]
-test_utils = []
-
 [dependencies]
 displaydoc = "0.2"
 hex = "0.4"
@@ -21,6 +18,7 @@ mc-transaction-core = { path = "../mobilecoin/transaction/core" }
 mc-transaction-extra = { path = "../mobilecoin/transaction/extra" }
 mc-transaction-types = { path = "../mobilecoin/transaction/types" }
 
+postage = "0.5"
 rand = "0.8" # build fails without this
 
 [dev-dependencies]
