+++ quote-book/src/lib.rs
@@ -7,11 +7,13 @@ mod error;
 mod in_memory_quote_book;
 mod quote;
 mod quote_book;
+mod msg;
 mod synchronized_quote_book;
 
 pub use basic_types::{Pair, QuoteId};
 pub use error::Error;
 pub use in_memory_quote_book::InMemoryQuoteBook;
+pub use msg::Msg;
 pub use synchronized_quote_book::SynchronizedQuoteBook;
 
 pub use quote::Quote;
