+++ quote-book/tests/synchronized_quote_book.rs
@@ -4,10 +4,11 @@ mod common;
 
 use std::time::Duration;
 
-use deqs_quote_book::{Error, InMemoryQuoteBook, QuoteBook, SynchronizedQuoteBook};
+use deqs_quote_book::{Error, InMemoryQuoteBook, QuoteBook, SynchronizedQuoteBook, Msg};
 use mc_blockchain_types::BlockVersion;
 use mc_crypto_ring_signature_signer::NoKeysRingSigner;
 use mc_ledger_db::{Ledger, LedgerDB};
+use postage::broadcast;
 use rand::{rngs::StdRng, SeedableRng};
 
 use mc_account_keys::AccountKey;
@@ -37,48 +38,54 @@ fn create_and_initialize_test_ledger() -> LedgerDB {
 #[test_with_logger]
 fn basic_happy_flow(logger: Logger) {
     let ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger, logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
     common::basic_happy_flow(&synchronized_quote_book);
 }
 
 #[test_with_logger]
 fn cannot_add_invalid_sci(logger: Logger) {
     let ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger, logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
     common::cannot_add_invalid_sci(&synchronized_quote_book);
 }
 
 #[test_with_logger]
 fn get_quotes_filtering_works(logger: Logger) {
     let ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger, logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
     common::get_quotes_filtering_works(&synchronized_quote_book);
 }
 
 #[test_with_logger]
 fn get_quote_ids_works(logger: Logger) {
     let ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger, logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
     common::get_quote_ids_works(&synchronized_quote_book);
 }
 
 #[test_with_logger]
 fn get_quote_by_id_works(logger: Logger) {
     let ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger, logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
     common::get_quote_by_id_works(&synchronized_quote_book);
 }
 
 #[test_with_logger]
 fn cannot_add_sci_with_key_image_in_ledger(logger: Logger) {
     let mut ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
 
     let pair = common::pair();
     let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
@@ -128,8 +135,9 @@ fn cannot_add_sci_with_key_image_in_ledger(logger: Logger) {
 #[test_with_logger]
 fn sci_that_are_added_to_ledger_are_removed_in_the_background(logger: Logger) {
     let mut ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
 
     let pair = common::pair();
     let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
@@ -183,8 +191,9 @@ fn cannot_add_sci_past_tombstone_block(logger: Logger) {
     let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
 
     let ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
     let internal_quote_book = InMemoryQuoteBook::default();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
 
     // Because the tombstone block is lower than the number blocks already in the
     // ledger, adding this sci should fail
@@ -236,9 +245,11 @@ fn sci_past_tombstone_block_get_removed_in_the_background(logger: Logger) {
     let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
 
     let mut ledger = create_and_initialize_test_ledger();
+    let (msg_bus_tx, _) = broadcast::channel::<Msg>(1000);
+    
     let internal_quote_book = InMemoryQuoteBook::default();
     let starting_blocks = ledger.num_blocks().unwrap();
-    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), logger);
+    let synchronized_quote_book = SynchronizedQuoteBook::new(internal_quote_book, ledger.clone(), msg_bus_tx, logger);
 
     // Because the tombstone block is lower than the number blocks already in the
     // ledger, adding this sci should fail
