+++ server/src/lib.rs
@@ -4,20 +4,12 @@ mod client_service;
 mod config;
 mod error;
 mod metrics;
-mod msg;
 mod p2p;
 mod server;
-mod synchronized_quote_book;
 
-#[cfg(any(test, feature = "test_utils"))]
-#[path="../../quote-book/tests/common/mod.rs"]
-mod test_common;
-
-pub use synchronized_quote_book::SynchronizedQuoteBook;
 pub use client_service::ClientService;
 pub use config::ServerConfig;
 pub use error::Error;
 pub use metrics::SVC_COUNTERS;
-pub use msg::Msg;
 pub use p2p::P2P;
 pub use server::Server;
