+++ server/src/server.rs
@@ -1,8 +1,8 @@
 // Copyright (c) 2023 MobileCoin Inc.
 
-use crate::{ClientService, Error, Msg};
+use crate::{ClientService, Error};
 use deqs_api::DeqsClientUri;
-use deqs_quote_book::QuoteBook;
+use deqs_quote_book::{QuoteBook, Msg};
 use futures::executor::block_on;
 use mc_common::logger::{log, Logger};
 use mc_util_grpc::ConnectionUriGrpcioServer;
